Add files via upload
Day 7

Hangman game
On the seventh day of my coding journey, I was able to code a hangman game. The task enabled me to think about the thought process of creating the game.
I was able to utilize what I had learnt on loops, importing of modules and I also learnt how to import another python file into my current file and call on a particular object in the python file.