Add files via upload
Day 8 (Functions with inputs)

1. feat: Paint area calculator
2. feat: prime checker
3 and 4.  game: Caesar cipher

On day 8, I learnt how to define functions and attribute inputs to it. I was able to sharpen my knowledge by coding a paint area calculator, a prime checker and a Caesar cipher game. 
I had to code the Caesar cipher twice because after coding the first game successfully without any notable errors, I discovered that instead of defining different functions for encrypting and decrypting the text, I could define a single to perform both operations, which I was able to do successfully and saved space by reducing the number of lines in my code while achieving the same result as the initial game code.

From this task, I began to reason how to make my code not just work, but for it to be neat, efficient and readable.